requirements.txt
setuptools
wheel
flake8
flake8-import-order
black
mypy
Pillow